requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.46.0
bottle==0.13.3
bottle-websocket==0.2.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
customtkinter==5.2.2
cython==3.1.2
darkdetect==0.8.0
dearpygui==2.0.0
eel==0.18.1
emoji==2.14.1
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
gputil==1.4.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
nuitka==2.7.12
numpy==2.3.0
opencv-python==4.11.0.86
ordered-set==4.1.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pluggy==1.6.0
psutil==7.0.0
pycparser==2.22
pycryptodome==3.23.0
pygments==2.19.1
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.3
pytest==8.4.0
pywin32==310
pywin32-ctypes==0.2.3
reedsolo==1.7.0
requests==2.32.3
setuptools==80.9.0
sv-ttk==2.6.0
typing-extensions==4.14.0
urllib3==2.4.0
winui==0.1.1
wmi==1.5.1
yt-dlp==2025.5.22
zope-event==5.0
zope-interface==7.2
zstandard==0.23.0
